+++ packages/jest-configurator/src/jest-configurator.js
@@ -77,6 +77,9 @@ export default (
     preset: "react-native",
     ...platformCode(platform),
     rootDir,
+    transform: {
+      "^.+\\.js$": path.resolve(__dirname, "source-loader.js")
+    },
     transformIgnorePatterns: [
       "node_modules/(?!(react-native|react-native-linear-gradient|react-native-iphone-x-helper|@times-components)/)"
     ],