+++ packages/author-head/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.25.3"></a>
+## [0.25.3](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.25.2...@times-components/author-head@0.25.3) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.25.2"></a>
+## [0.25.2](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.25.1...@times-components/author-head@0.25.2) (2018-04-25)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.25.1"></a>
+## [0.25.1](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.25.0...@times-components/author-head@0.25.1) (2018-04-24)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.25.0"></a>
+# [0.25.0](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.24.7...@times-components/author-head@0.25.0) (2018-04-24)
+
+
+### Features
+
+* image resizing ([#861](https://github.com/newsuk/times-components/issues/861)) ([73bb820](https://github.com/newsuk/times-components/commit/73bb820))
+
+
+
+
 <a name="0.24.7"></a>
 ## [0.24.7](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.24.6...@times-components/author-head@0.24.7) (2018-04-23)