+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.11.4"></a>
+## [0.11.4](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.11.3...@times-components/error-view@0.11.4) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
+<a name="0.11.3"></a>
+## [0.11.3](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.11.2...@times-components/error-view@0.11.3) (2018-04-25)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
+<a name="0.11.2"></a>
+## [0.11.2](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.11.1...@times-components/error-view@0.11.2) (2018-04-24)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
 <a name="0.11.1"></a>
 ## [0.11.1](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.11.0...@times-components/error-view@0.11.1) (2018-04-23)