+++ packages/pull-quote/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.5"></a>
+## [0.6.5](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@0.6.4...@times-components/pull-quote@0.6.5) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/pull-quote
+
+<a name="0.6.4"></a>
+## [0.6.4](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@0.6.3...@times-components/pull-quote@0.6.4) (2018-04-25)
+
+
+
+
+**Note:** Version bump only for package @times-components/pull-quote
+
+<a name="0.6.3"></a>
+## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@0.6.2...@times-components/pull-quote@0.6.3) (2018-04-24)
+
+
+
+
+**Note:** Version bump only for package @times-components/pull-quote
+
 <a name="0.6.2"></a>
 ## [0.6.2](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@0.6.1...@times-components/pull-quote@0.6.2) (2018-04-23)