+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.3"></a>
+## [0.3.3](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.3.2...@times-components/related-articles@0.3.3) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.3.2"></a>
+## [0.3.2](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.3.1...@times-components/related-articles@0.3.2) (2018-04-25)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.3.0...@times-components/related-articles@0.3.1) (2018-04-24)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.11...@times-components/related-articles@0.3.0) (2018-04-24)
+
+
+### Features
+
+* image resizing ([#861](https://github.com/newsuk/times-components/issues/861)) ([73bb820](https://github.com/newsuk/times-components/commit/73bb820))
+
+
+
+
+<a name="0.2.11"></a>
+## [0.2.11](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.10...@times-components/related-articles@0.2.11) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
 <a name="0.2.10"></a>
 ## [0.2.10](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.9...@times-components/related-articles@0.2.10) (2018-04-23)